use process.cwd() as the main directory in REPL.
This will prevent a fatal error, and instead allow a basic level of REPL support.  There are some limitations that might not be obvious:  changes to the process working directory will not be tracked, and the node-hotload module will sometimes fail silently if one saves a unsyntactic version of the code being watched, and require reloading from cache.